Add a userController file.
