+++ src/routes/userRoutes.js
@@ -1,86 +1,18 @@
 const { Router } = require('express');
-const jwt = require('jsonwebtoken');
-const User = require('../models/User');
 const auth = require('../middlewares/auth');
 const router = Router();
+const userController = require('../controllers/userController');
 
-router.post('/register', async (req, res) => {
-    try {
-        const user = new User(req.body);
-        await user.save();
-        res.status(201).json(user);
-    } catch (error) {
-        // console.log(error.message);
-        res.status(400).json({ msg: error.message });
-    }
-});
+router.post('/register', userController.registerUser);
 
-router.post('/login', async (req, res) => {
-    try {
-        const user = await User.findByCredentials(req.body.email, req.body.password);
-        const token = jwt.sign({ _id: user._id.toString() }, process.env.JWT_SECRET, { expiresIn: "7d" });
-        user.token = token;
+router.post('/login', userController.loginUser);
 
-        res.status(201).json(user);
-    } catch (error) {
-        res.status(401).json({ msg: error.message });
-    }
-});
+router.post('/logout', auth, userController.logoutUser);
 
-router.post('/logout', auth, async (req, res) => {
-    try {
-        const dbRes = await User.updateOne(
-            { _id: req.userId }, { $unset: { token: 1 } }
-        );
-        
-        if(!dbRes.acknowledged) {
-            res.json({msg: 'You are unauthorize!'});
-        }
+router.patch('/me', auth, userController.updateProfile);
 
-        res.json({ msg: 'You are loged out' });
-    } catch (error) {
-        res.status(400).json({ msg: error.message });
-    }
-});
+router.get('/me', auth, userController.getProfile);
 
-router.patch('/me', auth, async (req, res) => {
-    const validFields = ['username', 'email'];
-    const updates = Object.keys(req.body)
-    const fieldsAreValid = updates.every(field => validFields.includes(field));
-
-    try {
-        if (!fieldsAreValid) {
-            throw new Error('You entered invalid field');
-        }
-
-        const user = await User.findById(req.userId);
-        updates.forEach(update => user[update] = req.body[update]);
-
-        await user.save();
-        res.json({ msg: 'Profile successfully updated!' });
-    } catch (error) {
-        res.status(401).json({ msg: error.message });
-    }
-});
-
-router.get('/me', auth, async (req, res) => {
-    try {
-        const user = await User.findById({ _id: req.userId })
-        const { username, email, id } = user;
-        res.json({ username, email, id });
-    } catch (error) {
-        res.status(400).json({ msg: 'Something went wrong.' })
-    }
-});
-
-router.delete('/me', auth, async (req, res) => {
-    try {
-        await User.deleteOne({ _id: req.userId });
-
-        res.json({ msg: 'Your account is deleted successfully!' });
-    } catch (error) {
-        res.status(400).json({ msg: 'something went wrong!' });
-    }
-});
+router.delete('/me', auth, );
 
 module.exports = router;
